Unify page editing: persistent header with inline Save/Cancel
- Merged PageEditorView and PageReaderView into PageDetailView as a single
  unified view with a persistent header that stays fixed across reader/editor
  modes. Only the content area below the divider swaps.
- Save/Cancel buttons now live in the header (same row as Edit/Copy Path),
  matching the pattern established in IngestedFileDetailView.
- Aligned IngestedFileDetailView to the same pattern: Save/Cancel in header
  instead of bottom bar + toolbar button. Removed the toolbar Edit button.
- Removed PageEditorView.swift and PageReaderView.swift (dead code).
- PageDetailView markdown renders full-width (contentInset: false).
